+++ packages/gov-ui-kit/jest.config.js
@@ -4,10 +4,10 @@
 const config = {
     testEnvironment: 'jsdom',
     collectCoverageFrom: ['./src/**/*.{ts,tsx}'],
-    coveragePathIgnorePatterns: ['.d.ts', '.api.ts', 'index.ts', '.stories.tsx', './src/test/*'],
-    setupFilesAfterEnv: ['<rootDir>/src/test/setup.ts'],
+    coveragePathIgnorePatterns: ['.d.ts', '.api.ts', 'index.ts', '.stories.tsx', './src/core/test/*'],
+    setupFilesAfterEnv: ['<rootDir>/src/core/test/setup.ts'],
     transform: {
-        '^.+\\.svg$': '<rootDir>/src/test/svgTransform.js',
+        '^.+\\.svg$': '<rootDir>/src/core/test/svgTransform.js',
         '^.+\\.m?[tj]sx?$': 'ts-jest',
     },
     transformIgnorePatterns: ['node_modules/(?!(.*\\.mjs$|react-merge-refs))'],