+++ packages/gov-ui-kit/src/core/components/dialogs/dialog/dialogRoot/dialogRoot.tsx
@@ -1,6 +1,8 @@
 import { Content, Overlay, Portal, Root } from '@radix-ui/react-dialog';
 import classNames from 'classnames';
+import { AnimatePresence, motion } from 'framer-motion';
 import { type ComponentPropsWithoutRef, type ReactNode } from 'react';
+import { DialogUtils } from '../../dialogUtils';
 
 export interface IDialogRootProps extends ComponentPropsWithoutRef<'div'> {
     /**
@@ -68,23 +70,45 @@ export const DialogRoot: React.FC<IDialogRootProps> = (props) => {
 
     return (
         <Root {...rootProps}>
-            <Portal>
-                <Overlay className={classNames('fixed inset-0 bg-modal-overlay backdrop-blur-md', overlayClassName)} />
-                <Content
-                    className={classNames(
-                        'fixed inset-x-2 bottom-2 mx-auto max-h-[calc(100vh-80px)] lg:bottom-auto lg:top-[120px] lg:max-h-[calc(100vh-200px)]',
-                        'flex max-w-[480px] flex-col rounded-xl border border-neutral-100 bg-neutral-0 shadow-neutral-md md:min-w-[480px]',
-                        containerClassName,
-                    )}
-                    onCloseAutoFocus={onCloseAutoFocus}
-                    onEscapeKeyDown={onEscapeKeyDown}
-                    onInteractOutside={onInteractOutside}
-                    onOpenAutoFocus={onOpenAutoFocus}
-                    onPointerDownOutside={onPointerDownOutside}
-                >
-                    {children}
-                </Content>
-            </Portal>
+            <AnimatePresence>
+                {rootProps.open && (
+                    <Portal forceMount key="portal">
+                        <Overlay
+                            className={classNames('fixed inset-0 bg-modal-overlay backdrop-blur-md', overlayClassName)}
+                            asChild
+                        >
+                            <motion.div
+                                initial="closed"
+                                animate="open"
+                                exit="closed"
+                                variants={DialogUtils.overlayAnimationVariants}
+                            />
+                        </Overlay>
+                        <Content
+                            className={classNames(
+                                'fixed inset-x-2 bottom-2 mx-auto max-h-[calc(100vh-80px)] lg:bottom-auto lg:top-[120px] lg:max-h-[calc(100vh-200px)]',
+                                'flex max-w-[480px] flex-col rounded-xl border border-neutral-100 bg-neutral-0 shadow-neutral-md md:min-w-[480px]',
+                                containerClassName,
+                            )}
+                            onCloseAutoFocus={onCloseAutoFocus}
+                            onEscapeKeyDown={onEscapeKeyDown}
+                            onInteractOutside={onInteractOutside}
+                            onOpenAutoFocus={onOpenAutoFocus}
+                            onPointerDownOutside={onPointerDownOutside}
+                            asChild
+                        >
+                            <motion.div
+                                variants={DialogUtils.contentAnimationVariants}
+                                initial="closed"
+                                animate="open"
+                                exit="exit"
+                            >
+                                {children}
+                            </motion.div>
+                        </Content>
+                    </Portal>
+                )}
+            </AnimatePresence>
         </Root>
     );
 };