+++ packages/gov-ui-kit/src/modules/components/index.ts
@@ -1,2 +1,3 @@
 export * from './dao';
+export * from './member';
 export * from './odsModulesProvider';