+++ packages/gov-ui-kit/src/core/components/dialogs/dialogAlert/dialogAlertRoot/dialogAlertRoot.tsx
@@ -82,7 +82,11 @@ export const DialogAlertRoot: React.FC<IDialogAlertRootProps> = (props) => {
                 {rootProps.open && (
                     <Portal forceMount key="portal">
                         <Overlay
-                            className={classNames('fixed inset-0 bg-modal-overlay backdrop-blur-md', overlayClassName)}
+                            className={classNames(
+                                'fixed inset-0 bg-modal-overlay backdrop-blur-md',
+                                'z-[var(--ods-dialog-alert-overlay-z-index)]',
+                                overlayClassName,
+                            )}
                             asChild
                         >
                             <motion.div
@@ -96,6 +100,7 @@ export const DialogAlertRoot: React.FC<IDialogAlertRootProps> = (props) => {
                             className={classNames(
                                 'fixed inset-x-2 bottom-2 mx-auto max-h-[calc(100vh-80px)] lg:bottom-auto lg:top-[120px] lg:max-h-[calc(100vh-200px)]',
                                 'flex max-w-[480px] flex-col rounded-xl border border-neutral-100 bg-neutral-0 shadow-neutral-md md:min-w-[480px]',
+                                'z-[var(--ods-dialog-alert-content-z-index)]',
                                 containerClassName,
                             )}
                             onCloseAutoFocus={onCloseAutoFocus}