+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Update library structure to support module components
+-   Update documentation about how to install the library
+
 ## [1.0.18] - 2024-02-29
 
 ### Fixed