+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 -   Implement animations for `Dialog` and `DialogAlert` components
+-   Implement `DaoAvatar` module component
 
 ### Changed